requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
httpx==0.27.2
python-dotenv==1.0.1
groq==1.0.0